Implement sf::WindowBase and CreateVulkanSurface
