+++ src/SFML.Window/Window.cs
@@ -8,38 +8,10 @@ namespace SFML.Window
 {
     ////////////////////////////////////////////////////////////
     /// <summary>
-    /// Enumeration of window creation styles
+    /// Window that serves as a target for OpenGL rendering
     /// </summary>
     ////////////////////////////////////////////////////////////
-    [Flags]
-    public enum Styles
-    {
-        /// <summary>No border / title bar (this flag and all others are mutually exclusive)</summary>
-        None = 0,
-
-        /// <summary>Title bar + fixed border</summary>
-        Titlebar = 1 << 0,
-
-        /// <summary>Titlebar + resizable border + maximize button</summary>
-        Resize = 1 << 1,
-
-        /// <summary>Titlebar + close button</summary>
-        Close = 1 << 2,
-
-        /// <summary>Fullscreen mode (this flag and all others are mutually exclusive))</summary>
-        Fullscreen = 1 << 3,
-
-        /// <summary>Default window style (titlebar + resize + close)</summary>
-        Default = Titlebar | Resize | Close
-    }
-
-    ////////////////////////////////////////////////////////////
-    /// <summary>
-    /// Window is a rendering window ; it can create a new window
-    /// or connect to an existing one
-    /// </summary>
-    ////////////////////////////////////////////////////////////
-    public class Window : ObjectBase
+    public class Window : WindowBase
     {
         ////////////////////////////////////////////////////////////
         /// <summary>
@@ -121,7 +93,7 @@ namespace SFML.Window
         /// </summary>
         /// <returns>True if the window is opened</returns>
         ////////////////////////////////////////////////////////////
-        public virtual bool IsOpen
+        public override bool IsOpen
         {
             get { return sfWindow_isOpen(CPointer); }
         }
@@ -133,7 +105,7 @@ namespace SFML.Window
         /// Create to recreate the window
         /// </summary>
         ////////////////////////////////////////////////////////////
-        public virtual void Close()
+        public override void Close()
         {
             sfWindow_close(CPointer);
         }
@@ -163,7 +135,7 @@ namespace SFML.Window
         /// Position of the window
         /// </summary>
         ////////////////////////////////////////////////////////////
-        public virtual Vector2i Position
+        public override Vector2i Position
         {
             get { return sfWindow_getPosition(CPointer); }
             set { sfWindow_setPosition(CPointer, value); }
@@ -174,7 +146,7 @@ namespace SFML.Window
         /// Size of the rendering region of the window
         /// </summary>
         ////////////////////////////////////////////////////////////
-        public virtual Vector2u Size
+        public override Vector2u Size
         {
             get { return sfWindow_getSize(CPointer); }
             set { sfWindow_setSize(CPointer, value); }
@@ -186,7 +158,7 @@ namespace SFML.Window
         /// </summary>
         /// <param name="title">New title</param>
         ////////////////////////////////////////////////////////////
-        public virtual void SetTitle(string title)
+        public override void SetTitle(string title)
         {
             // Copy the title to a null-terminated UTF-32 byte array
             byte[] titleAsUtf32 = Encoding.UTF32.GetBytes(title + '\0');
@@ -208,7 +180,7 @@ namespace SFML.Window
         /// <param name="height">Icon's height, in pixels</param>
         /// <param name="pixels">Array of pixels, format must be RGBA 32 bits</param>
         ////////////////////////////////////////////////////////////
-        public virtual void SetIcon(uint width, uint height, byte[] pixels)
+        public override void SetIcon(uint width, uint height, byte[] pixels)
         {
             unsafe
             {
@@ -225,7 +197,7 @@ namespace SFML.Window
         /// </summary>
         /// <param name="visible">True to show the window, false to hide it</param>
         ////////////////////////////////////////////////////////////
-        public virtual void SetVisible(bool visible)
+        public override void SetVisible(bool visible)
         {
             sfWindow_setVisible(CPointer, visible);
         }
@@ -236,7 +208,7 @@ namespace SFML.Window
         /// </summary>
         /// <param name="show">True to show, false to hide</param>
         ////////////////////////////////////////////////////////////
-        public virtual void SetMouseCursorVisible(bool show)
+        public override void SetMouseCursorVisible(bool show)
         {
             sfWindow_setMouseCursorVisible(CPointer, show);
         }
@@ -256,7 +228,7 @@ namespace SFML.Window
         /// cursor).
         /// </remarks>
         ////////////////////////////////////////////////////////////
-        public virtual void SetMouseCursorGrabbed(bool grabbed)
+        public override void SetMouseCursorGrabbed(bool grabbed)
         {
             sfWindow_setMouseCursorGrabbed(CPointer, grabbed);
         }
@@ -267,7 +239,7 @@ namespace SFML.Window
         /// </summary>
         /// <param name="cursor">Native system cursor type to display</param>
         ////////////////////////////////////////////////////////////
-        public virtual void SetMouseCursor(Cursor cursor)
+        public override void SetMouseCursor(Cursor cursor)
         {
             sfWindow_setMouseCursor(CPointer, cursor.CPointer);
         }
@@ -290,7 +262,7 @@ namespace SFML.Window
         /// </summary>
         /// <param name="enable">True to enable, false to disable</param>
         ////////////////////////////////////////////////////////////
-        public virtual void SetKeyRepeatEnabled(bool enable)
+        public override void SetKeyRepeatEnabled(bool enable)
         {
             sfWindow_setKeyRepeatEnabled(CPointer, enable);
         }
@@ -338,7 +310,7 @@ namespace SFML.Window
         /// </summary>
         /// <param name="threshold">New threshold, in range [0, 100]</param>
         ////////////////////////////////////////////////////////////
-        public virtual void SetJoystickThreshold(float threshold)
+        public override void SetJoystickThreshold(float threshold)
         {
             sfWindow_setJoystickThreshold(CPointer, threshold);
         }
@@ -348,38 +320,9 @@ namespace SFML.Window
         /// OS-specific handle of the window
         /// </summary>
         ////////////////////////////////////////////////////////////
-        public virtual IntPtr SystemHandle
+        public override IntPtr SystemHandle
         {
             get { return sfWindow_getSystemHandle(CPointer); }
-        }
-
-        ////////////////////////////////////////////////////////////
-        /// <summary>
-        /// Wait for a new event and dispatch it to the corresponding
-        /// event handler
-        /// </summary>
-        ////////////////////////////////////////////////////////////
-        public void WaitAndDispatchEvents()
-        {
-            Event e;
-            if (WaitEvent(out e))
-            {
-                CallEventHandler(e);
-            }
-        }
-
-        ////////////////////////////////////////////////////////////
-        /// <summary>
-        /// Call the event handlers for each pending event
-        /// </summary>
-        ////////////////////////////////////////////////////////////
-        public void DispatchEvents()
-        {
-            Event e;
-            while (PollEvent(out e))
-            {
-                CallEventHandler(e);
-            }
         }
 
         ////////////////////////////////////////////////////////////
@@ -388,7 +331,7 @@ namespace SFML.Window
         /// foreground window
         /// </summary>
         ////////////////////////////////////////////////////////////
-        public virtual void RequestFocus()
+        public override void RequestFocus()
         {
             sfWindow_requestFocus(CPointer);
         }
@@ -399,7 +342,7 @@ namespace SFML.Window
         /// </summary>
         /// <returns>True if the window has focus, false otherwise</returns>
         ////////////////////////////////////////////////////////////
-        public virtual bool HasFocus()
+        public override bool HasFocus()
         {
             return sfWindow_hasFocus(CPointer);
         }
@@ -426,7 +369,7 @@ namespace SFML.Window
         /// <param name="dummy">Internal hack :)</param>
         ////////////////////////////////////////////////////////////
         protected Window(IntPtr cPointer, int dummy) :
-            base(cPointer)
+            base(cPointer, 0)
         {
             // TODO : find a cleaner way of separating this constructor from Window(IntPtr handle)
         }
@@ -438,7 +381,7 @@ namespace SFML.Window
         /// <param name="eventToFill">Variable to fill with the raw pointer to the event structure</param>
         /// <returns>True if there was an event, false otherwise</returns>
         ////////////////////////////////////////////////////////////
-        protected virtual bool PollEvent(out Event eventToFill)
+        protected override bool PollEvent(out Event eventToFill)
         {
             return sfWindow_pollEvent(CPointer, out eventToFill);
         }
@@ -450,49 +393,9 @@ namespace SFML.Window
         /// <param name="eventToFill">Variable to fill with the raw pointer to the event structure</param>
         /// <returns>False if any error occured</returns>
         ////////////////////////////////////////////////////////////
-        protected virtual bool WaitEvent(out Event eventToFill)
+        protected override bool WaitEvent(out Event eventToFill)
         {
             return sfWindow_waitEvent(CPointer, out eventToFill);
-        }
-
-        ////////////////////////////////////////////////////////////
-        /// <summary>
-        /// Internal function to get the mouse position relative to the window.
-        /// This function is protected because it is called by another class of
-        /// another module, it is not meant to be called by users.
-        /// </summary>
-        /// <returns>Relative mouse position</returns>
-        ////////////////////////////////////////////////////////////
-        protected internal virtual Vector2i InternalGetMousePosition()
-        {
-            return sfMouse_getPosition(CPointer);
-        }
-
-        ////////////////////////////////////////////////////////////
-        /// <summary>
-        /// Internal function to set the mouse position relative to the window.
-        /// This function is protected because it is called by another class of
-        /// another module, it is not meant to be called by users.
-        /// </summary>
-        /// <param name="position">Relative mouse position</param>
-        ////////////////////////////////////////////////////////////
-        protected internal virtual void InternalSetMousePosition(Vector2i position)
-        {
-            sfMouse_setPosition(position, CPointer);
-        }
-
-        ////////////////////////////////////////////////////////////
-        /// <summary>
-        /// Internal function to get the touch position relative to the window.
-        /// This function is protected because it is called by another class of
-        /// another module, it is not meant to be called by users.
-        /// </summary>
-        /// <param name="Finger">Finger index</param>
-        /// <returns>Relative touch position</returns>
-        ////////////////////////////////////////////////////////////
-        protected internal virtual Vector2i InternalGetTouchPosition(uint Finger)
-        {
-            return sfTouch_getPosition(Finger, CPointer);
         }
 
         ////////////////////////////////////////////////////////////
@@ -505,279 +408,6 @@ namespace SFML.Window
         {
             sfWindow_destroy(CPointer);
         }
-
-        ////////////////////////////////////////////////////////////
-        /// <summary>
-        /// Call the event handler for the given event
-        /// </summary>
-        /// <param name="e">Event to dispatch</param>
-        ////////////////////////////////////////////////////////////
-        private void CallEventHandler(Event e)
-        {
-            switch (e.Type)
-            {
-                case EventType.Closed:
-                    if (Closed != null)
-                    {
-                        Closed(this, EventArgs.Empty);
-                    }
-
-                    break;
-
-                case EventType.GainedFocus:
-                    if (GainedFocus != null)
-                    {
-                        GainedFocus(this, EventArgs.Empty);
-                    }
-
-                    break;
-
-                case EventType.JoystickButtonPressed:
-                    if (JoystickButtonPressed != null)
-                    {
-                        JoystickButtonPressed(this, new JoystickButtonEventArgs(e.JoystickButton));
-                    }
-
-                    break;
-
-                case EventType.JoystickButtonReleased:
-                    if (JoystickButtonReleased != null)
-                    {
-                        JoystickButtonReleased(this, new JoystickButtonEventArgs(e.JoystickButton));
-                    }
-
-                    break;
-
-                case EventType.JoystickMoved:
-                    if (JoystickMoved != null)
-                    {
-                        JoystickMoved(this, new JoystickMoveEventArgs(e.JoystickMove));
-                    }
-
-                    break;
-
-                case EventType.JoystickConnected:
-                    if (JoystickConnected != null)
-                    {
-                        JoystickConnected(this, new JoystickConnectEventArgs(e.JoystickConnect));
-                    }
-
-                    break;
-
-                case EventType.JoystickDisconnected:
-                    if (JoystickDisconnected != null)
-                    {
-                        JoystickDisconnected(this, new JoystickConnectEventArgs(e.JoystickConnect));
-                    }
-
-                    break;
-
-                case EventType.KeyPressed:
-                    if (KeyPressed != null)
-                    {
-                        KeyPressed(this, new KeyEventArgs(e.Key));
-                    }
-
-                    break;
-
-                case EventType.KeyReleased:
-                    if (KeyReleased != null)
-                    {
-                        KeyReleased(this, new KeyEventArgs(e.Key));
-                    }
-
-                    break;
-
-                case EventType.LostFocus:
-                    if (LostFocus != null)
-                    {
-                        LostFocus(this, EventArgs.Empty);
-                    }
-
-                    break;
-
-                case EventType.MouseButtonPressed:
-                    if (MouseButtonPressed != null)
-                    {
-                        MouseButtonPressed(this, new MouseButtonEventArgs(e.MouseButton));
-                    }
-
-                    break;
-
-                case EventType.MouseButtonReleased:
-                    if (MouseButtonReleased != null)
-                    {
-                        MouseButtonReleased(this, new MouseButtonEventArgs(e.MouseButton));
-                    }
-
-                    break;
-
-                case EventType.MouseEntered:
-                    if (MouseEntered != null)
-                    {
-                        MouseEntered(this, EventArgs.Empty);
-                    }
-
-                    break;
-
-                case EventType.MouseLeft:
-                    if (MouseLeft != null)
-                    {
-                        MouseLeft(this, EventArgs.Empty);
-                    }
-
-                    break;
-
-                case EventType.MouseMoved:
-                    if (MouseMoved != null)
-                    {
-                        MouseMoved(this, new MouseMoveEventArgs(e.MouseMove));
-                    }
-
-                    break;
-
-// Disable CS0618 (Obselete Warning).  This Event will be removed in SFML.NET 3.0, but should remain supported until then.
-#pragma warning disable CS0618
-                case EventType.MouseWheelMoved:
-                    if (MouseWheelMoved != null)
-                    {
-                        MouseWheelMoved(this, new MouseWheelEventArgs(e.MouseWheel));
-                    }
-
-                    break;
-// restore CS0618
-#pragma warning restore CS0618
-
-                case EventType.MouseWheelScrolled:
-                    if (MouseWheelScrolled != null)
-                    {
-                        MouseWheelScrolled(this, new MouseWheelScrollEventArgs(e.MouseWheelScroll));
-                    }
-
-                    break;
-
-                case EventType.Resized:
-                    if (Resized != null)
-                    {
-                        Resized(this, new SizeEventArgs(e.Size));
-                    }
-
-                    break;
-
-                case EventType.TextEntered:
-                    if (TextEntered != null)
-                    {
-                        TextEntered(this, new TextEventArgs(e.Text));
-                    }
-
-                    break;
-
-                case EventType.TouchBegan:
-                    if (TouchBegan != null)
-                    {
-                        TouchBegan(this, new TouchEventArgs(e.Touch));
-                    }
-
-                    break;
-
-                case EventType.TouchMoved:
-                    if (TouchMoved != null)
-                    {
-                        TouchMoved(this, new TouchEventArgs(e.Touch));
-                    }
-
-                    break;
-
-                case EventType.TouchEnded:
-                    if (TouchEnded != null)
-                    {
-                        TouchEnded(this, new TouchEventArgs(e.Touch));
-                    }
-
-                    break;
-
-                case EventType.SensorChanged:
-                    if (SensorChanged != null)
-                    {
-                        SensorChanged(this, new SensorEventArgs(e.Sensor));
-                    }
-
-                    break;
-
-                default:
-                    break;
-            }
-        }
-
-        /// <summary>Event handler for the Closed event</summary>
-        public event EventHandler Closed = null;
-
-        /// <summary>Event handler for the Resized event</summary>
-        public event EventHandler<SizeEventArgs> Resized = null;
-
-        /// <summary>Event handler for the LostFocus event</summary>
-        public event EventHandler LostFocus = null;
-
-        /// <summary>Event handler for the GainedFocus event</summary>
-        public event EventHandler GainedFocus = null;
-
-        /// <summary>Event handler for the TextEntered event</summary>
-        public event EventHandler<TextEventArgs> TextEntered = null;
-
-        /// <summary>Event handler for the KeyPressed event</summary>
-        public event EventHandler<KeyEventArgs> KeyPressed = null;
-
-        /// <summary>Event handler for the KeyReleased event</summary>
-        public event EventHandler<KeyEventArgs> KeyReleased = null;
-
-        /// <summary>Event handler for the MouseWheelMoved event</summary>
-        [Obsolete("Use MouseWheelScrolled")]
-        public event EventHandler<MouseWheelEventArgs> MouseWheelMoved = null;
-
-        /// <summary>Event handler for the MouseWheelScrolled event</summary>
-        public event EventHandler<MouseWheelScrollEventArgs> MouseWheelScrolled = null;
-
-        /// <summary>Event handler for the MouseButtonPressed event</summary>
-        public event EventHandler<MouseButtonEventArgs> MouseButtonPressed = null;
-
-        /// <summary>Event handler for the MouseButtonReleased event</summary>
-        public event EventHandler<MouseButtonEventArgs> MouseButtonReleased = null;
-
-        /// <summary>Event handler for the MouseMoved event</summary>
-        public event EventHandler<MouseMoveEventArgs> MouseMoved = null;
-
-        /// <summary>Event handler for the MouseEntered event</summary>
-        public event EventHandler MouseEntered = null;
-
-        /// <summary>Event handler for the MouseLeft event</summary>
-        public event EventHandler MouseLeft = null;
-
-        /// <summary>Event handler for the JoystickButtonPressed event</summary>
-        public event EventHandler<JoystickButtonEventArgs> JoystickButtonPressed = null;
-
-        /// <summary>Event handler for the JoystickButtonReleased event</summary>
-        public event EventHandler<JoystickButtonEventArgs> JoystickButtonReleased = null;
-
-        /// <summary>Event handler for the JoystickMoved event</summary>
-        public event EventHandler<JoystickMoveEventArgs> JoystickMoved = null;
-
-        /// <summary>Event handler for the JoystickConnected event</summary>
-        public event EventHandler<JoystickConnectEventArgs> JoystickConnected = null;
-
-        /// <summary>Event handler for the JoystickDisconnected event</summary>
-        public event EventHandler<JoystickConnectEventArgs> JoystickDisconnected = null;
-
-        /// <summary>Event handler for the TouchBegan event</summary>
-        public event EventHandler<TouchEventArgs> TouchBegan = null;
-
-        /// <summary>Event handler for the TouchMoved event</summary>
-        public event EventHandler<TouchEventArgs> TouchMoved = null;
-
-        /// <summary>Event handler for the TouchEnded event</summary>
-        public event EventHandler<TouchEventArgs> TouchEnded = null;
-
-        /// <summary>Event handler for the SensorChanged event</summary>
-        public event EventHandler<SensorEventArgs> SensorChanged = null;
 
         #region Imports
         [DllImport(CSFML.window, CallingConvention = CallingConvention.Cdecl), SuppressUnmanagedCodeSecurity]
@@ -869,15 +499,6 @@ namespace SFML.Window
 
         [DllImport(CSFML.window, CallingConvention = CallingConvention.Cdecl), SuppressUnmanagedCodeSecurity]
         private static extern bool sfWindow_hasFocus(IntPtr CPointer);
-
-        [DllImport(CSFML.window, CallingConvention = CallingConvention.Cdecl), SuppressUnmanagedCodeSecurity]
-        private static extern Vector2i sfMouse_getPosition(IntPtr CPointer);
-
-        [DllImport(CSFML.window, CallingConvention = CallingConvention.Cdecl), SuppressUnmanagedCodeSecurity]
-        private static extern void sfMouse_setPosition(Vector2i position, IntPtr CPointer);
-
-        [DllImport(CSFML.window, CallingConvention = CallingConvention.Cdecl), SuppressUnmanagedCodeSecurity]
-        private static extern Vector2i sfTouch_getPosition(uint Finger, IntPtr RelativeTo);
         #endregion
     }
 }
